Update to parsing routine
Removed full iteration through the cites dictionary to replace payload
values with Jinja2 variables and it now directly updates without another
iteration construct. In addition fix a problem with content not mapping
to replacement payloads, for instance content maybe the same but the
payload signature maybe different as a result of for instance markup
surrounding identical content.